+++ common/src/main/java/com/lowdragmc/photon/command/EntityEffectCommand.java
@@ -43,12 +43,14 @@ public class EntityEffectCommand extends EffectCommand {
                         .executes(c -> execute(c, 0))
                         .then(Commands.argument("offset", Vec3Argument.vec3(false))
                                 .executes(c -> execute(c, 1))
-                                .then(Commands.argument("delay", IntegerArgumentType.integer(0))
+                                .then(Commands.argument("rotation", Vec3Argument.vec3(false))
                                         .executes(c -> execute(c, 2))
-                                        .then(Commands.argument("force death", BoolArgumentType.bool())
+                                        .then(Commands.argument("delay", IntegerArgumentType.integer(0))
                                                 .executes(c -> execute(c, 3))
-                                                .then(Commands.argument("allow multi", BoolArgumentType.bool())
-                                                        .executes(c -> execute(c, 4)))))));
+                                                .then(Commands.argument("force death", BoolArgumentType.bool())
+                                                        .executes(c -> execute(c, 4))
+                                                        .then(Commands.argument("allow multi", BoolArgumentType.bool())
+                                                                .executes(c -> execute(c, 5))))))));
     }
 
     private static int execute(CommandContext<CommandSourceStack> context, int feature) throws CommandSyntaxException {
@@ -59,12 +61,15 @@ public class EntityEffectCommand extends EffectCommand {
             command.setOffset(Vec3Argument.getVec3(context, "offset"));
         }
         if (feature >= 2) {
-            command.setDelay(IntegerArgumentType.getInteger(context, "delay"));
+            command.setRotation(Vec3Argument.getVec3(context, "rotation"));
         }
         if (feature >= 3) {
-            command.setForcedDeath(BoolArgumentType.getBool(context, "force death"));
+            command.setDelay(IntegerArgumentType.getInteger(context, "delay"));
         }
         if (feature >= 4) {
+            command.setForcedDeath(BoolArgumentType.getBool(context, "force death"));
+        }
+        if (feature >= 5) {
             command.setAllowMulti(BoolArgumentType.getBool(context, "allow multi"));
         }
         PhotonNetworking.NETWORK.sendToAll(command);
@@ -100,6 +105,7 @@ public class EntityEffectCommand extends EffectCommand {
                 if (entity != null) {
                     var effect = new EntityEffect(fx, level, entity);
                     effect.setOffset(offset.x, offset.y, offset.z);
+                    effect.setRotation(rotation.x, offset.y, offset.z);
                     effect.setDelay(delay);
                     effect.setForcedDeath(forcedDeath);
                     effect.setAllowMulti(allowMulti);